requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
click==7.1.2
dj-database-url==0.5.0
Django==3.1.1
django-common-helpers==0.9.2
django-crontab==0.7.1
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
gunicorn==20.0.4
mypy-extensions==0.4.3
pathspec==0.8.0
psycopg2-binary==2.8.6
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0